+++ src/ponyfill.js
@@ -1,4 +1,5 @@
 /**
+ * @package
  * @returns {(fetch|import("undici/types/fetch").fetch)} The fetch function
  */
 function pickFetch() {
