+++ src/types/text.js
